leftover suggestions from review of PR #11787
